fix: use minor currency units (int64/BIGINT) for balance

+++ payment-service/internal/models/account.go
@@ -1,6 +1,7 @@
 package models
 
+// Account balances are stored in minor currency units (e.g. kopecks).
 type Account struct {
 	UserID  int64
-	Balance float64
+	Balance int64
 }
